+++ bot/__init__.py
@@ -1,0 +1,8 @@
+from .bot import DiscordBot
+from .bot_event_handler import DiscordBotEventHandler
+
+
+__all__ = [
+    'DiscordBot',
+    'DiscordBotEventHandler',
+]
