+++ bot/commands/message_bot_channel.pyi
@@ -1,6 +1,6 @@
 from discord import Message
 
-from ..discord_bot import DiscordBot
+from ..bot import DiscordBot
 
 
 async def message_bot_channel(message: Message, bot: DiscordBot): ...
