+++ bot/commands/ping.py
