+++ bot/commands/restart.py
