+++ bot/commands/restart.pyi
@@ -1,6 +1,6 @@
 from discord import Message
 
-from ..discord_bot import DiscordBot
+from ..bot import DiscordBot
 
 
 async def message_restart(message: Message, args_str: str, bot: DiscordBot): ...
