+++ bot/commands/shutdown.py
