+++ bot/moduels/__init__.py
@@ -1,0 +1,10 @@
+from .module import Module
+from .pics_sending_module import PicsSendingModule
+from .pics_suggestion_module import PicsSuggestionModule
+
+
+__all__ = [
+    'Module',
+    'PicsSendingModule',
+    'PicsSuggestionModule',
+]
