+++ bot/moduels/pics_suggestion_module.pyi
@@ -1,0 +1,7 @@
+from .module import Module
+from ..bot import DiscordBot
+
+
+class PicsSuggestionModule(Module):
+
+    def __init__(self, category_name: str, bot: DiscordBot): ...
