+++ bot/utils/__init__.py
@@ -1,0 +1,10 @@
+from .config import parse_config
+from .logger import init_logger
+from .utils import get_pics_path_list
+
+
+__all__ = [
+    'parse_config',
+    'init_logger',
+    'get_pics_path_list',
+]
