+++ bot/utils/utils.pyi
