--- utils/__init__.py
@@ -1,18 +1,0 @@
-from .config import parse_config
-from .discord_bot import DiscordBot
-from .discord_event_handler import DiscordEventHandler
-from .logger import init_logger
-from .pics_sending_manager import PicsSendingManager
-from .pics_suggestion_manager import PicsSuggestionManager
-from .utils import get_pics_path_list
-
-
-__all__ = [
-    'parse_config',
-    'DiscordBot',
-    'DiscordEventHandler',
-    'init_logger',
-    'PicsSendingManager',
-    'PicsSuggestionManager',
-    'get_pics_path_list',
-]
